DRF-TODO/requirements.txt
asgiref==3.6.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Faker==12.0.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
sqlparse==0.4.3
psycopg2-binary==2.9.5
mixer~=7.2.2


